requirements.txt
flask
pytesseract
pillow
joblib
web3.storage
pdf2image